nested loop

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -16,7 +16,7 @@
 
 int main(void)
 {
-	int fib[n];
+	long long fib[n];
 	int i;
 
 	fib[0] = 1;
